requirements.txt
PySoundFile==0.9.0.post1
torchaudio==0.10.1
torchvision==0.11.2
torchsummary==1.5.1

